% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Models.R
\name{compare_models}
\alias{compare_models}
\title{compare_models()}
\usage{
compare_models(
  null.vars,
  added.vars,
  time,
  status,
  data,
  weighted = TRUE,
  dept.censor = FALSE
)
}
\arguments{
\item{null.vars}{Vector of variable names for the null model}

\item{added.vars}{Vector of variable names to be added to the null model to yield the full model}

\item{time}{Character string. Name of the time-to-event variable}

\item{status}{Character string. Name of the event status variable}

\item{data}{Data frame containing time, status, and the variables specified in the null and full models}

\item{weighted}{Boolean, default TRUE}

\item{dept.censor}{Boolean, default FALSE}
}
\value{
Generates side-by-side plots comparing concordance gains by
traditional C-index and proposed new Smooth C-index, as well as:
 * C.df a data frame of c-indices
}
\description{
Allows the user to easily compare the gains in concordance, using both
traditional C-index and the proposed new Smooth C-index, between two nested
survival models.
}
