% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_C_SmoothC.R
\name{find_C_smoothC}
\alias{find_C_smoothC}
\title{find_C_smoothC()}
\usage{
find_C_smoothC(object, weighted = TRUE, dept.censor = FALSE)
}
\arguments{
\item{object}{a fitted coxph-type model object.}

\item{weighted}{Boolean, default TRUE. Use IPCW weights to find C-indices?}

\item{dept.censor}{Boolean, default FALSE. Is censoring is dependent on predictors?}
}
\value{
c.vals List of two numeric values:
 * C : traditional concordance index value
 * smooth.C : proposed new c-statistic; coefficient of logit regression for standardized risk differences
}
\description{
Calculates both the traditional concordance index (C) and the proposed smooth C-statistic
for a fitted survival model.
}
